backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-google-genai>=1.0.7
langgraph
pydantic
requests
google-generativeai>=0.7.0
faiss-cpu
pymupdf
python-multipart
langchain-core
duckduckgo-search
